Extract theme config
