+++ docs/.vuepress/config.js
@@ -1,8 +1,5 @@
 const { resolve } = require('path')
-const footnotes = require('markdown-it-footnote')
-const implicitFigures = require('markdown-it-implicit-figures')
-const slugify = require('../../theme/slugify')
-const preprocessMarkdown = resolve(__dirname, '../../theme/preprocessMarkdown')
+const themeConfig = require('../../theme/config')
 
 const title = 'Bitcoin Dev Kit Documentation'
 const baseUrl = 'https://bitcoindevkit.org'
@@ -10,17 +7,7 @@ const githubUrl = 'https://github.com/bitcoindevkit'
 const discordUrl = 'https://discord.gg/dstn4dQ'
 const twitterUrl = 'https://twitter.com/intent/follow?screen_name=bitcoindevkit'
 const themeColor = '#ffffff'
-const pageSuffix = '/'
-const info = { name: title }
-const extractDescription = text => {
-  if (!text) return
-  const paragraph = text.match(/^[A-Za-z].*(?:\n[A-Za-z].*)*/m)
-  return paragraph ? paragraph.toString().replace(/[\*\_\(\)\[\]]/g, '') : null
-}
-const sitemap = {
-  hostname: baseUrl,
-  exclude: ['/404.html']
-}
+
 const docsSidebar = [
   {
     title: 'Documentation',
@@ -67,113 +54,12 @@ const blogSidebar = [
 module.exports = {
   title,
   description: 'The Bitcoin Dev Kit (BDK) project (originally called Magical Bitcoin 🧙) aims to build a collection of tools and libraries that are designed to be a solid foundation for cross platform Bitcoin wallets, along with a fully working reference implementation wallet called Magical Bitcoin.',
-  head: [
-    ['meta', { name: 'viewport', content: 'width=device-width,initial-scale=1.0' }],
-    ['link', { rel: 'preload', href: '/fonts/manrope-400.woff2', as: 'font', crossorigin: true }],
-    ['link', { rel: 'preload', href: '/fonts/manrope-600.woff2', as: 'font', crossorigin: true }],
-    ['link', { rel: 'preload', href: '/fonts/ibm-plex-mono-400.woff2', as: 'font', crossorigin: true }],
-    ['link', { rel: 'apple-touch-icon', sizes: '180x180', href: '/img/favicon/apple-touch-icon.png' }],
-    ['link', { rel: 'manifest', href: '/site.webmanifest' }],
-    ['link', { rel: 'stylesheet', href: '/css/variables.css' }],
-    ['link', { name: 'msapplication-config', content: '/browserconfig.xml' }],
-    ['link', { name: 'msapplication-TileColor', content: themeColor }],
-    ['link', { name: 'theme-color', content: themeColor }]
-  ],
-  chainWebpack (config) {
-    config.module
-      .rule('md')
-      .test(/\.md$/)
-      .use(preprocessMarkdown)
-        .loader(preprocessMarkdown)
-        .end()
-  },
-  plugins: [
-    ['seo', {
-      siteTitle: (_, $site) => $site.title,
-      title: $page => $page.title,
-      description: $page => $page.frontmatter.description || extractDescription($page._strippedContent),
-      author: (_, $site) => info,
-      tags: $page => ($page.frontmatter.tags || ['Bitcoin', 'Bitcoin Dev Kit', 'BDK']),
-      twitterCard: _ => 'summary',
-      type: $page => 'article',
-      url: (_, $site, path) => `${baseUrl}${path.replace('.html', pageSuffix)}`,
-      image: ($page, $site) => `${baseUrl}/card.png`
-    }],
-    ['clean-urls', {
-      normalSuffix: pageSuffix,
-      indexSuffix: pageSuffix,
-      notFoundPath: '/404.html',
-    }],
-    ['code-copy', {
-      color: '#8F979E',
-      backgroundTransition: false,
-      staticIcon: true
-    }],
-    ['mermaidjs'],
-    ['sitemap', sitemap],
-    ['tabs', {
-      tabsAttributes: {
-        options: { useUrlFragment: false }
-      }
-    }],
-    ['@vuepress/medium-zoom'],
-    ['@vuepress/blog', {
-      sitemap,
-      directories: [
-        {
-          id: 'blog',
-          dirname: '_blog',
-          path: '/blog/',
-          itemPermalink: '/blog/:slug',
-          pagination: {
-            lengthPerPage: 10,
-            getPaginationPageTitle(pageNumber) {
-              return `Page ${pageNumber}`
-            }
-          }
-        },
-      ],
-      frontmatters: [
-        {
-          id: 'tags',
-          keys: ['tags'],
-          path: '/blog/tags/',
-          title: '',
-          frontmatter: {
-            title: 'Tags'
-          },
-          pagination: {
-            getPaginationPageTitle(pageNumber, key) {
-              return `${capitalize(key)} - Page ${pageNumber}`
-            }
-          }
-        },
-        {
-          id: 'author',
-          keys: ['author', 'authors'],
-          path: '/blog/author/',
-          title: '',
-          frontmatter: {
-            title: 'Authors'
-          },
-          pagination: {
-            getPaginationPageTitle(pageNumber, key) {
-              return `${key} - Page ${pageNumber}`
-            }
-          }
-        },
-      ],
-    }]
-  ],
-  markdown: {
-    extendMarkdown(md) {
-      md.use(footnotes)
-      md.use(implicitFigures)
-    },
-    pageSuffix,
-    slugify
-  },
   theme: resolve(__dirname, '../../theme'),
+  ...themeConfig({
+    baseUrl,
+    title,
+    themeColor
+  }),
   themeConfig: {
     domain: baseUrl,
     logo: '/img/logo.svg',
