+++ package.json
@@ -32,11 +32,11 @@
     "@vuepress/plugin-back-to-top": "1.8.2",
     "@vuepress/plugin-blog": "1.9.4",
     "@vuepress/plugin-medium-zoom": "1.8.2",
+    "@vuepress/plugin-search": "1.8.2",
     "broken-link-checker": "0.7.8",
-    "markdown-it-footnote": "3.0.3",
-    "markdown-it-implicit-figures": "0.10.0",
     "serve": "12.0.0",
     "start-server-and-test": "1.14.0",
+    "vue-tabs-component": "1.5.0",
     "vuepress": "1.8.2",
     "vuepress-plugin-clean-urls": "1.1.2",
     "vuepress-plugin-code-copy": "1.0.6",
@@ -44,8 +44,5 @@
     "vuepress-plugin-seo": "0.1.4",
     "vuepress-plugin-sitemap": "2.3.1",
     "vuepress-plugin-tabs": "0.3.0"
-  },
-  "dependencies": {
-    "vue-tabs-component": "1.5.0"
   }
 }
