+++ theme/package.json
@@ -19,6 +19,20 @@
   "author": "Dennis Reimann",
   "main": "index.js",
   "dependencies": {
-    "@vuepress/plugin-search": "1.8.2"
+    "@vuepress/plugin-back-to-top": "1.8.2",
+    "@vuepress/plugin-blog": "1.9.4",
+    "@vuepress/plugin-medium-zoom": "1.8.2",
+    "@vuepress/plugin-search": "1.8.2",
+    "@vuepress/theme-default": "1.8.2",
+    "lodash": "4.17.21",
+    "markdown-it-footnote": "3.0.3",
+    "markdown-it-implicit-figures": "0.10.0",
+    "vue-tabs-component": "1.5.0",
+    "vuepress-plugin-clean-urls": "1.1.2",
+    "vuepress-plugin-code-copy": "1.0.6",
+    "vuepress-plugin-mermaidjs": "^1.8.1",
+    "vuepress-plugin-seo": "0.1.4",
+    "vuepress-plugin-sitemap": "2.3.1",
+    "vuepress-plugin-tabs": "0.3.0"
   }
 }
